Remove addresses forms
