+++ app/forms/waste_exemptions_engine/contact_address_lookup_form.rb
@@ -2,7 +2,19 @@
 
 module WasteExemptionsEngine
   class ContactAddressLookupForm < AddressLookupForm
-    include ContactAddressForm
 
+    private
+
+    def existing_postcode
+      @transient_registration.temp_contact_postcode
+    end
+
+    def existing_address
+      @transient_registration.contact_address
+    end
+
+    def address_type
+      TransientAddress.address_types[:contact]
+    end
   end
 end
