+++ app/forms/waste_exemptions_engine/operator_address_lookup_form.rb
@@ -2,6 +2,20 @@
 
 module WasteExemptionsEngine
   class OperatorAddressLookupForm < AddressLookupForm
-    include OperatorAddressForm
+    delegate :business_type, to: :transient_registration
+
+    private
+
+    def existing_postcode
+      @transient_registration.temp_operator_postcode
+    end
+
+    def existing_address
+      @transient_registration.operator_address
+    end
+
+    def address_type
+      TransientAddress.address_types[:operator]
+    end
   end
 end
