+++ app/forms/waste_exemptions_engine/site_address_lookup_form.rb
@@ -2,7 +2,19 @@
 
 module WasteExemptionsEngine
   class SiteAddressLookupForm < AddressLookupForm
-    include SiteAddressForm
 
+    private
+
+    def existing_postcode
+      @transient_registration.temp_site_postcode
+    end
+
+    def existing_address
+      @transient_registration.site_address
+    end
+
+    def address_type
+      TransientAddress.address_types[:site]
+    end
   end
 end
